Enable read queries from access node standbys
In high availability setups, a streaming replica of the access node can
be configured. Since it's a hot standby users should be able to run
read-only queries from this replica and should be able to query the
distributed hypertables from it.

To allow this, we need to disable the "timescaledb.enable_2pc" GUC on
the standby.

Additionally, the internal code which sets up a connection from
the standby access node to the data nodes now marks those sessions
explicitly as READ ONLY ones. This ensures that any functions or any
activity which tries to make modifications on the data nodes errors
out.