Refactor to combine concat files command
Refactor the code to combine two different ways to create a command to
concatenate files. The current two functions `concatenate_files` and
`cat_files` are refactored to create a single `add_concat_command`
working in a similar way to `add_custom_command` but setting up a
concatenation command with dependencies on the source files so that the
target is re-generated if the source files are out of date.

To be able to use this with the generated script files, the
`generate_script` and `generate_downgrade_script` have a new parameter
`RESULT_FILE` that takes the name of a variable that will contain the
resulting script file. This is then added to the list of
install/update/downgrade scripts for the custom target so that the
files are regenerated if the input changes.