Fix targetlist for cagg views
If the resource names for entries in the targetlist for the views for
the user view does not match the attribute names in the tuple
descriptor for the result tuple an error will be generated. This commit
fixes this by setting the targetlist resource names to the attribute
name.

Fixes #3051
Fixes #3405